Refactor DecodeQuestion to take the number of questions as parameter
DNS specification allows more than one query in DNS packet. However, more than
one question in packet is not supported by most of resolvers thus it is
unlikely that such a packet is really captured.

Changed the implementation to return qname, qtype from last record on
question section. This ensures that if such packet is received, we are able to
parse other sections following the question.