Implement relabeling functionality in the Loki logger.
This allows the user to control the labels on the entries sent to Loki,
similar to how promtail allows it:
https://grafana.com/docs/loki/latest/clients/promtail/configuration/#relabel_configs.
This should grealy improve query performance by allowing the user to
configure labels to filter their most relevant log entries.

There is no change to the default behavior of the logger with regards to
the `identity` and `job` labels. But they can now be adjusted by the
relabel-config as well.
Log entries can also be dropped if all labels are removed.